+++ src/beam/llm/__init__.py
@@ -1,0 +1,1 @@
+from llm import beam_llm
