+++ src/beam/llm/beam_llm.py
@@ -1,136 +1,17 @@
-from .processor import Processor
-from .path import beam_key
+import json
+import re
+import string
+import time
+from functools import partial
+from typing import Optional, Any, Dict, List, Mapping
+
+import numpy as np
 import pandas as pd
 
-import json
-import numpy as np
-import string
-
-from functools import partial
-from .utils import get_edit_ratio, normalize_host, BeamURL, retry, parse_text_to_protocol
-import openai
-from typing import Any, List, Mapping, Optional, Dict
-
-from langchain.callbacks.manager import CallbackManagerForLLMRun
-from langchain.llms.base import LLM
-from pydantic import BaseModel, Field, PrivateAttr
-from transformers.pipelines import Conversation
-import requests
-import uuid
-import time
-from argparse import Namespace
-from .logger import beam_logger as logger
-
-
-def simulate_openai_chat(model=None, **kwargs):
-    llm = beam_llm(model) if type(model) == str else model
-    return llm.chat_completion(**kwargs).openai_format
-
-
-def simulate_openai_completion(model=None, **kwargs):
-    llm = beam_llm(model) if type(model) == str else model
-    return llm.completion(**kwargs).openai_format
-
-
-openai_simulator = Namespace(ChatCompletion=Namespace(create=simulate_openai_chat),
-                             Completion=Namespace(create=simulate_openai_completion))
-
-
-def estimate_tokens(s):
-    # Pattern to capture:
-    # 1. Any single punctuation or whitespace character.
-    # 2. Sequences of up to 4 alphabetic characters.
-    # 3. Sequences of up to 3 numeric characters.
-    pattern = f" ?[{re.escape(string.whitespace + string.punctuation)}]| ?[A-Za-z]{{1,4}}| ?\d{{1,3}}"
-
-    matches = re.findall(pattern, s)
-    token_count = len(matches)
-
-    # For the remaining unmatched characters in the string (those that aren't part of the recognized patterns),
-    # we count each character as a separate token.
-    unmatched_characters = re.sub(pattern, "", s)
-    token_count += len(unmatched_characters)
-
-    return token_count
-
-
-class LLMResponse:
-    def __init__(self, response, prompt, llm, chat=False, stream=False, retrials=3, sleep=1, **kwargs):
-        self.response = response
-        self.prompt = prompt
-        self.retrials = retrials
-        self.sleep = sleep
-        self.llm = llm
-        self.id = f'beamllm-{uuid.uuid4()}'
-        self.model = llm.model
-        self.created = int(time.time())
-        self.chat = chat
-        self.object = "chat.completion" if chat else "text_completion"
-        self.stream = stream
-        assert self.verify(), "Response is not valid"
-
-    def __iter__(self):
-        if not self.stream:
-            yield self
-        else:
-            for r in self.response:
-                yield LLMResponse(r, self.prompt, self.llm, self.chat, self.stream)
-
-    def verify(self):
-        return self.llm.verify_response(self)
-
-    @property
-    def text(self):
-        return self.llm.extract_text(self)
-
-    @property
-    def openai_format(self):
-        return self.llm.openai_format(self)
-
-    def _protocol(self, text, protocol='json'):
-        try:
-            return parse_text_to_protocol(text, protocol=protocol)
-        except:
-            retry_protocol = (retry(retrials=self.retrials, sleep=self.sleep,  logger=logger, name=f"fix-{protocol} with {self.model}")
-                              (self.llm.fix_protocol))
-            return retry_protocol(text, protocol=protocol)
-
-    @property
-    def json(self):
-        json_text = self.llm.extract_text(self)
-        json_text = json_text.replace('False', 'false')
-        json_text = json_text.replace('True', 'true')
-        return self._protocol(json_text, protocol='json')
-
-    @property
-    def html(self):
-        text = self.llm.extract_text(self)
-        return self._protocol(text, protocol='html')
-
-    @property
-    def xml(self):
-        text = self.llm.extract_text(self)
-        return self._protocol(text, protocol='xml')
-
-    @property
-    def csv(self):
-        text = self.llm.extract_text(self)
-        return self._protocol(text, protocol='csv')
-
-    @property
-    def yaml(self):
-        text = self.llm.extract_text(self)
-        # text = re.search('yaml\n([\s\S]*?)\n', text).group(1)
-        return self._protocol(text, protocol='yaml')
-
-    @property
-    def toml(self):
-        text = self.llm.extract_text(self)
-        return self._protocol(text, protocol='toml')
-
-    @property
-    def choices(self):
-        return self.llm.extract_choices(self.response)
+from beam import beam_logger as logger
+from beam.llm.llm_response import LLMResponse
+from beam.processor import Processor
+from beam.utils import BeamURL, parse_text_to_protocol, get_edit_ratio
 
 
 class BeamLLM(LLM, Processor):
@@ -842,657 +723,19 @@ class BeamLLM(LLM, Processor):
         pass
 
 
-class OpenAIBase(BeamLLM):
-
-    api_key: Optional[str] = Field(None)
-    api_base: Optional[str] = Field(None)
-    organization: Optional[str] = Field(None)
-
-    def __init__(self, api_key=None, api_base=None, organization=None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.api_key = api_key
-        self.api_base = api_base
-        self.organization = organization
-
-    def update_usage(self, response):
-
-        if 'usage' in response:
-            response = response['usage']
-
-            self.usage["prompt_tokens"] += response["prompt_tokens"]
-            self.usage["completion_tokens"] += response["completion_tokens"]
-            self.usage["total_tokens"] += response["prompt_tokens"] + response["completion_tokens"]
-
-    def sync_openai(self):
-        openai.api_key = self.api_key
-        openai.api_base = self.api_base
-        openai.organization = self.organization
-
-    def _chat_completion(self, **kwargs):
-        self.sync_openai()
-        return openai.ChatCompletion.create(model=self.model, **kwargs)
-
-    def _completion(self, **kwargs):
-        self.sync_openai()
-        return openai.Completion.create(engine=self.model, **kwargs)
-
-    def verify_response(self, res):
-        res = res.response
-        finish_reason = res.choices[0].finish_reason
-        if finish_reason != 'stop':
-            logger.warning(f"finish_reason is {finish_reason}")
-        return True
-
-    def extract_text(self, res):
-
-        stream = res.stream
-        res = res.response
-
-        if not self.is_chat:
-            res = res.choices[0].text
-        else:
-            if not stream:
-                res = res.choices[0].message.content
-            else:
-                res = res.choices[0].delta.content
-        return res
-
-    def openai_format(self, res):
-
-        res = res.response
-        return res
-
-
-class FCConversationLLM(BeamLLM):
-
-    _ma: Any = PrivateAttr()
-    _conv: Any = PrivateAttr()
-
-    def __init__(self, *args, model=None, model_adapter=None, **kwargs):
-
-        super().__init__(*args, model=model, **kwargs)
-
-        if model_adapter is not None:
-            model = model_adapter
-
-        from fastchat.model.model_adapter import get_model_adapter
-        self._ma = get_model_adapter(model)
-        self._conv = self._ma.get_default_conv_template('       ')
-
-    @property
-    def stop_sequence(self):
-        return self._conv.stop_str
-
-    def get_prompt(self, messages):
-
-        conv = self._ma.get_default_conv_template('       ')
-
-        for m in messages:
-
-            if 'system' in m:
-
-                role = m['system_name'] if 'system_name' in m else 'system'
-                content = m['system']
-
-            else:
-
-                if m['role'] == 'user':
-                    role = conv.roles[0]
-                else:
-                    role = conv.roles[1]
-
-                content = m['content']
-
-            conv.append_message(role, content)
-
-        conv.append_message(self._conv.roles[1], None)
-
-        return conv.get_prompt()
-
-    @property
-    def is_chat(self):
-        return True
-
-    @property
-    def is_completions(self):
-        return True
-
-
-class TGILLM(FCConversationLLM):
-
-    _info: Any = PrivateAttr()
-    _client: Any = PrivateAttr()
-
-    def __init__(self, hostname=None, port=None, *args, **kwargs):
-
-        api_base = f"http://{normalize_host(hostname, port)}"
-
-        from text_generation import Client
-        self._client = Client(api_base)
-
-        req = requests.get(f"{api_base}/info")
-        self._info = json.loads(req.text)
-
-        kwargs['model'] = self._info['model_id']
-        kwargs['scheme'] = 'tgi'
-        super().__init__(*args, **kwargs)
-
-    def update_usage(self, response):
-
-        self.usage["prompt_tokens"] += 0
-        self.usage["completion_tokens"] += response.details.generated_tokens
-        self.usage["total_tokens"] += 0 + response.details.generated_tokens
-
-    def openai_format(self, res):
-
-        text = self.extract_text(res)
-
-        if res.chat:
-            choice = {
-                      "finish_reason": "stop",
-                      "index": 0,
-                      "message": {
-                        "content": text,
-                        "role": "assistant"
-                      }
-                    }
-        else:
-            choice = {
-                "finish_reason": "stop",
-                "index": 0,
-                "logprobs": res.response.details.tokens,
-                "text": text
-            }
-
-        res = {
-                  "choices": [
-                    choice
-                  ],
-                  "created": res.created,
-                  "id": res.id,
-                  "model": res.model,
-                  "object": res.object,
-                  "usage": {
-                    "completion_tokens": res.response.details.generated_tokens,
-                    "prompt_tokens": 0,
-                    "total_tokens": res.response.details.generated_tokens
-                  }
-                }
-
-        return res
-
-    @property
-    def is_chat(self):
-        return False
-
-    @property
-    def is_completions(self):
-        return True
-
-    def process_kwargs(self, prompt, **kwargs):
-
-        processed_kwargs = {}
-
-        max_tokens = kwargs.pop('max_tokens', None)
-        max_new_tokens = None
-
-        if max_tokens is not None:
-            max_new_tokens = max_tokens - self.len_function(prompt)
-
-        max_new_tokens = kwargs.pop('max_new_tokens', max_new_tokens)
-        if max_new_tokens is not None:
-            processed_kwargs['max_new_tokens'] = max_new_tokens
-
-        temperature = kwargs.pop('temperature', None)
-        if temperature is not None:
-            processed_kwargs['temperature'] = temperature
-
-        top_p = kwargs.pop('top_p', None)
-        if top_p is not None and 0 < top_p < 1:
-            processed_kwargs['top_p'] = top_p
-
-        best_of = kwargs.pop('n', None)
-        if best_of is not None and best_of > 1:
-            processed_kwargs['best_of'] = best_of
-
-        stop_sequences = kwargs.pop('stop', None)
-        if stop_sequences is None:
-            stop_sequences = []
-        elif type(stop_sequences) is str:
-            stop_sequences = [stop_sequences]
-
-        if self.stop_sequence is not None:
-            stop_sequences.append(self.stop_sequence)
-
-        if len(stop_sequences) > 0:
-            processed_kwargs['stop_sequences'] = stop_sequences
-
-        if 'top_k' in kwargs:
-            processed_kwargs['top_k'] = kwargs.pop('top_k')
-
-        if 'truncate' in kwargs:
-            processed_kwargs['truncate'] = kwargs.pop('truncate')
-
-        if 'typical_p' in kwargs:
-            processed_kwargs['typical_p'] = kwargs.pop('typical_p')
-
-        if 'watermark' in kwargs:
-            processed_kwargs['watermark'] = kwargs.pop('watermark')
-
-        if 'repetition_penalty' in kwargs:
-            processed_kwargs['repetition_penalty'] = kwargs.pop('repetition_penalty')
-
-        decoder_input_details = kwargs.pop('logprobs', None)
-        if decoder_input_details is not None:
-            processed_kwargs['decoder_input_details'] = decoder_input_details
-
-        if 'seed' in kwargs:
-            processed_kwargs['seed'] = kwargs.pop('seed')
-
-        do_sample = None
-        if temperature is not None and temperature > 0:
-            do_sample = True
-        do_sample = kwargs.pop('do_sample', do_sample)
-
-        if do_sample is not None:
-            processed_kwargs['do_sample'] = do_sample
-
-        return processed_kwargs
-
-    def _completion(self, prompt=None, **kwargs):
-
-        prompt = self.get_prompt([{'role': 'user', 'content': prompt}])
-        generate_kwargs = self.process_kwargs(prompt, **kwargs)
-        return self._client.generate(prompt, **generate_kwargs)
-
-    def _chat_completion(self, messages=None, **kwargs):
-
-        prompt = self.get_prompt(messages)
-        generate_kwargs = self.process_kwargs(prompt, **kwargs)
-
-        return self._client.generate(prompt, **generate_kwargs)
-
-    def extract_text(self, res):
-
-        res = res.response
-        text = res.generated_text
-        text = text.rstrip(self.stop_sequence)
-        return text
-
-
-class FastChatLLM(OpenAIBase):
-
-    def __init__(self, model=None, hostname=None, port=None, *args, **kwargs):
-
-        api_base = f"http://{normalize_host(hostname, port)}/v1"
-        api_key = "EMPTY"  # Not support yet
-        organization = "EMPTY"  # Not support yet
-
-        kwargs['scheme'] = 'fastchat'
-        super().__init__(*args, api_key=api_key, api_base=api_base, organization=organization, model=model, **kwargs)
-
-        self.model = model
-
-    @property
-    def is_chat(self):
-        return True
-
-
-class FastAPILLM(FCConversationLLM):
-
-    model: Optional[str] = Field(None)
-    hostname: Optional[str] = Field(None)
-    headers: Optional[dict] = Field(None)
-    consumer: Optional[str] = Field(None)
-    protocol: Optional[str] = Field(None)
-    _models: Any = PrivateAttr()
-
-    def __init__(self, *args, model=None, hostname=None, port=None, username=None, protocol='https', **kwargs):
-
-        kwargs['scheme'] = 'fastapi'
-        super().__init__(*args, model=model, **kwargs)
-
-        self.consumer = username
-        self.hostname = normalize_host(hostname, port)
-        self._models = None
-        self.headers = {'Content-Type': 'application/json'}
-        self.protocol = protocol
-
-        from requests.packages.urllib3.exceptions import InsecureRequestWarning
-        # Suppress only the single InsecureRequestWarning from urllib3
-        requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-
-    @property
-    def models(self):
-        if self._models is None:
-            res = requests.get(f"{self.protocol}://{self.hostname}/models", headers=self.headers, verify=False)
-            self._models = res.json()
-        return self._models
-
-    @property
-    def is_chat(self):
-        return True
-
-    @property
-    def is_completions(self):
-        return True
-
-    def process_kwargs(self, prompt, **kwargs):
-
-        kwargs_processed = {}
-
-        max_tokens = kwargs.pop('max_tokens', None)
-        max_new_tokens = None
-
-        if max_tokens is not None:
-            max_new_tokens = max_tokens - self.len_function(prompt)
-
-        max_new_tokens = kwargs.pop('max_new_tokens', max_new_tokens)
-        if max_new_tokens is not None:
-            kwargs_processed['max_new_tokens'] = max_new_tokens
-
-        temperature = kwargs.pop('temperature', None)
-        if temperature is not None:
-            kwargs_processed['temp'] = temperature
-
-        return kwargs_processed
-
-    def _chat_completion(self, messages=None, **kwargs):
-
-        d = {}
-        d['model_name'] = self.model
-        d['consumer'] = self.consumer
-        d['input'] = self.get_prompt(messages)
-        d['hyper_params'] = self.process_kwargs(d['input'], **kwargs)
-
-        res = requests.post(f"{self.protocol}://{self.hostname}/predict/once", headers=self.headers, json=d,
-                            verify=False)
-        return res.json()
-
-    def _completion(self, prompt=None, **kwargs):
-
-        d = {}
-        d['model_name'] = self.model
-        d['consumer'] = self.consumer
-        d['input'] = self.get_prompt([{'role': 'user', 'content': prompt}])
-        d['hyper_params'] = self.process_kwargs(d['input'], **kwargs)
-
-        res = requests.post(f"{self.protocol}://{self.hostname}/predict/once", headers=self.headers, json=d,
-                            verify=False)
-        return res.json()
-
-    def verify_response(self, res):
-        try:
-
-            if not res.response['is_done']:
-                logger.warning(f"Model {self.model} is_done=False.")
-
-            assert 'res' in res.response, f"Response does not contain 'res' key"
-
-        except Exception as e:
-            logger.error(f"Error in response: {res.response}")
-            raise e
-        return True
-
-    def extract_text(self, res):
-        return res.response['res']
-
-
-class HuggingFaceLLM(BeamLLM):
-    config: Any
-    tokenizer: Any
-    net: Any
-    pipline_kwargs: Any
-    input_device: Optional[str] = Field(None)
-    eos_pattern: Optional[str] = Field(None)
-    _text_generation_pipeline: Any = PrivateAttr()
-    _conversational_pipeline: Any = PrivateAttr()
-
-    def __init__(self, model, tokenizer=None, dtype=None, chat=False, input_device=None, compile=True, *args,
-                 model_kwargs=None,
-                 config_kwargs=None, pipline_kwargs=None, text_generation_kwargs=None, conversational_kwargs=None,
-                 eos_pattern=None, **kwargs):
-
-        kwargs['scheme'] = 'huggingface'
-        kwargs['model'] = model
-        super().__init__(*args, **kwargs)
-
-        from transformers import AutoTokenizer, AutoModelForCausalLM, AutoConfig
-        import transformers
-
-        transformers.logging.set_verbosity_error()
-
-        if model_kwargs is None:
-            model_kwargs = {}
-
-        if config_kwargs is None:
-            config_kwargs = {}
-
-        if pipline_kwargs is None:
-            pipline_kwargs = {}
-
-        if text_generation_kwargs is None:
-            text_generation_kwargs = {}
-
-        if conversational_kwargs is None:
-            conversational_kwargs = {}
-
-        self.pipline_kwargs = pipline_kwargs
-
-        self.input_device = input_device
-        self.eos_pattern = eos_pattern
-
-        self.config = AutoConfig.from_pretrained(model, trust_remote_code=True, **config_kwargs)
-        tokenizer_name = tokenizer or model
-        self.tokenizer = AutoTokenizer.from_pretrained(tokenizer_name, trust_remote_code=True)
-
-        self.net = AutoModelForCausalLM.from_pretrained(model, trust_remote_code=True,
-                                                        config=self.config, **model_kwargs)
-
-        if compile:
-            import torch
-            self.net = torch.compile(self.net)
-
-        self._text_generation_pipeline = transformers.pipeline('text-generation', model=self.net,
-                                                               tokenizer=self.tokenizer, device=self.input_device,
-                                                               return_full_text=False, **text_generation_kwargs)
-
-        self._conversational_pipeline = transformers.pipeline('conversational', model=self.net,
-                                                              tokenizer=self.tokenizer, device=self.input_device,
-                                                              **conversational_kwargs)
-
-    def extract_text(self, res):
-
-        res = res.response
-        if type(res) is list:
-            res = res[0]
-
-        if type(res) is Conversation:
-            res = res.generated_responses[-1]
-        else:
-            res = res['generated_text']
-
-        if self.eos_pattern:
-            res = res.split(self.eos_pattern)[0]
-
-        return res
-
-    @property
-    def is_chat(self):
-        return True
-
-    @property
-    def is_completions(self):
-        return True
-
-    def _completion(self, prompt=None, **kwargs):
-
-        # pipeline = transformers.pipeline('text-generation', model=self.model,
-        #                                  tokenizer=self.tokenizer, device=self.input_device, return_full_text=False)
-
-        res = self._text_generation_pipeline(prompt, pad_token_id=self._text_generation_pipeline.tokenizer.eos_token_id,
-                                             **self.pipline_kwargs)
-
-        return res
-
-    def _chat_completion(self, **kwargs):
-
-        # pipeline = transformers.pipeline('conversational', model=self.model,
-        #                                  tokenizer=self.tokenizer, device=self.input_device)
-
-        return self._conversational_pipeline(self.conversation,
-                                             pad_token_id=self._conversational_pipeline.tokenizer.eos_token_id,
-                                             **self.pipline_kwargs)
-
-
-class OpenAI(OpenAIBase):
-
-    _models: Any = PrivateAttr()
-
-    def __init__(self, model='gpt-3.5-turbo', api_key=None, organization=None, *args, **kwargs):
-
-        api_key = beam_key('OPENAI_API_KEY', api_key)
-
-        kwargs['scheme'] = 'openai'
-        super().__init__(api_key=api_key, api_base='https://api.openai.com/v1',
-                         organization=organization, *args, **kwargs)
-
-        self.model = model
-        self._models = None
-
-    @property
-    def is_chat(self):
-        chat_models = ['gpt-4', 'gpt-4-0314', 'gpt-4-32k', 'gpt-4-32k-0314', 'gpt-3.5-turbo', 'gpt-3.5-turbo-0301']
-        if any([m in self.model for m in chat_models]):
-            return True
-        return False
-
-    def file_list(self):
-        return openai.File.list()
-
-    def build_dataset(self, data=None, question=None, answer=None, path=None) -> object:
-        """
-        Build a dataset for training a model
-        :param data: dataframe with prompt and completion columns
-        :param question: list of questions
-        :param answer: list of answers
-        :param path: path to save the dataset
-        :return: path to the dataset
-        """
-        if data is None:
-            data = pd.DataFrame(data={'prompt': question, 'completion': answer})
-
-        records = data.to_dict(orient='records')
-
-        if path is None:
-            print('No path provided, using default path: dataset.jsonl')
-            path = 'dataset.jsonl'
-
-        # Open a file for writing
-        with open(path, 'w') as outfile:
-            # Write each data item to the file as a separate line
-            for item in records:
-                json.dump(item, outfile)
-                outfile.write('\n')
-
-        return path
-
-    def retrieve(self, model=None):
-        if model is None:
-            model = self.model
-        return openai.Engine.retrieve(id=model)
-
-    @property
-    def models(self):
-        if self._models is None:
-            models = openai.Model.list()
-            models = {m.id: m for m in models.data}
-            self._models = models
-        return self._models
-
-    def embedding(self, text, model=None):
-        if model is None:
-            model = self.model
-        response = openai.Engine(model).embedding(input=text, model=model)
-        embedding = np.array(response.data[1]['embedding'])
-        return embedding
-
-
-def beam_llm(url, username=None, hostname=None, port=None, api_key=None, **kwargs):
-
-    if type(url) != str:
-        return url
-
-    url = BeamURL.from_string(url)
-
-    if url.hostname is not None:
-        hostname = url.hostname
-
-    if url.port is not None:
-        port = url.port
-
-    if url.username is not None:
-        username = url.username
-
-    query = url.query
-    for k, v in query.items():
-        kwargs[k] = v
-
-    if api_key is None and 'api_key' in kwargs:
-        api_key = kwargs.pop('api_key')
-
-    model = url.path
-    model = model.strip('/')
-    if not model:
-        model = None
-
-    if url.protocol == 'openai':
-
-        api_key = beam_key('OPENAI_API_KEY', api_key)
-        return OpenAI(model=model, api_key=api_key, **kwargs)
-
-    elif url.protocol == 'fastchat':
-        return FastChatLLM(model=model, hostname=hostname, port=port, **kwargs)
-
-    elif url.protocol == 'huggingface':
-        return HuggingFaceLLM(model=model, **kwargs)
-
-    elif url.protocol == 'fastapi':
-        return FastAPILLM(model=model, hostname=hostname, port=port, username=username, **kwargs)
-
-    elif url.protocol == 'tgi':
-        return TGILLM(model=model, hostname=hostname, port=port, username=username, **kwargs)
-
-    else:
-        raise NotImplementedError
-
-
-import re
-
-
-def text_splitter(text, separators=["\n\n", ". ", " "], chunk_size=100, length_function=None):
-    if length_function is None:
-        length_function = lambda x: int(1.5 * len(re.findall(r'\w+', x)))
-
-    s = separators[0]
-    open_chunks = text.split(s)
-    closed_chunks = []
-    next_chunk = ''
-
-    for c in open_chunks:
-        if length_function(c) > chunk_size:
-            if len(next_chunk) > 0:
-                closed_chunks.append(next_chunk)
-            closed_chunks.extend(text_splitter(c, separators[1:], chunk_size, length_function))
-            next_chunk = closed_chunks.pop()
-        elif length_function(next_chunk) + length_function(c) > chunk_size:
-            closed_chunks.append(next_chunk)
-            next_chunk = c
-        else:
-            next_chunk = f"{next_chunk}{s}{c}"
-
-    closed_chunks.append(next_chunk)
-
-    return closed_chunks
-
+def estimate_tokens(s):
+    # Pattern to capture:
+    # 1. Any single punctuation or whitespace character.
+    # 2. Sequences of up to 4 alphabetic characters.
+    # 3. Sequences of up to 3 numeric characters.
+    pattern = f" ?[{re.escape(string.whitespace + string.punctuation)}]| ?[A-Za-z]{{1,4}}| ?\d{{1,3}}"
+
+    matches = re.findall(pattern, s)
+    token_count = len(matches)
+
+    # For the remaining unmatched characters in the string (those that aren't part of the recognized patterns),
+    # we count each character as a separate token.
+    unmatched_characters = re.sub(pattern, "", s)
+    token_count += len(unmatched_characters)
+
+    return token_count
